+++ mcp_tools/param_discovery/arjun.py
@@ -1,0 +1,40 @@
+# mcp_tools/param_discovery/arjun.py
+
+from typing import Dict, Any
+
+def register_arjun_tool(mcp, hexstrike_client, logger):
+    @mcp.tool()
+    def arjun_parameter_discovery(url: str, method: str = "GET", wordlist: str = "",
+                                 delay: int = 0, threads: int = 25, stable: bool = False,
+                                 additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute Arjun for HTTP parameter discovery with enhanced logging.
+
+        Args:
+            url: The target URL
+            method: HTTP method to use
+            wordlist: Custom wordlist file
+            delay: Delay between requests
+            threads: Number of threads
+            stable: Use stable mode
+            additional_args: Additional Arjun arguments
+
+        Returns:
+            HTTP parameter discovery results
+        """
+        data = {
+            "url": url,
+            "method": method,
+            "wordlist": wordlist,
+            "delay": delay,
+            "threads": threads,
+            "stable": stable,
+            "additional_args": additional_args
+        }
+        logger.info(f"🎯 Starting Arjun parameter discovery: {url}")
+        result = hexstrike_client.safe_post("api/tools/arjun", data)
+        if result.get("success"):
+            logger.info(f"✅ Arjun parameter discovery completed for {url}")
+        else:
+            logger.error(f"❌ Arjun parameter discovery failed for {url}")
+        return result
