+++ mcp_tools/param_discovery/paramspider.py
@@ -1,0 +1,36 @@
+# mcp_tools/param_discovery/paramspider.py
+
+from typing import Dict, Any
+
+def register_paramspider_tool(mcp, hexstrike_client, logger):
+    @mcp.tool()
+    def paramspider_mining(domain: str, level: int = 2,
+                          exclude: str = "png,jpg,gif,jpeg,swf,woff,svg,pdf,css,ico",
+                          output: str = "", additional_args: str = "") -> Dict[str, Any]:
+        """
+        Execute ParamSpider for parameter mining from web archives with enhanced logging.
+
+        Args:
+            domain: The target domain
+            level: Mining level depth
+            exclude: File extensions to exclude
+            output: Output file path
+            additional_args: Additional ParamSpider arguments
+
+        Returns:
+            Parameter mining results from web archives
+        """
+        data = {
+            "domain": domain,
+            "level": level,
+            "exclude": exclude,
+            "output": output,
+            "additional_args": additional_args
+        }
+        logger.info(f"🕷️  Starting ParamSpider mining: {domain}")
+        result = hexstrike_client.safe_post("api/tools/paramspider", data)
+        if result.get("success"):
+            logger.info(f"✅ ParamSpider mining completed for {domain}")
+        else:
+            logger.error(f"❌ ParamSpider mining failed for {domain}")
+        return result
