+++ mcp_tools/web_app_security/enhanced_web_app_security.py
@@ -4,77 +4,6 @@ from typing import Dict, Any, Optional
 
 def register_enhanced_web_app_security_tools(mcp, hexstrike_client, logger, HexStrikeColors):
 
-
-
-
-    @mcp.tool()
-    def arjun_parameter_discovery(url: str, method: str = "GET", wordlist: str = "",
-                                 delay: int = 0, threads: int = 25, stable: bool = False,
-                                 additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute Arjun for HTTP parameter discovery with enhanced logging.
-
-        Args:
-            url: The target URL
-            method: HTTP method to use
-            wordlist: Custom wordlist file
-            delay: Delay between requests
-            threads: Number of threads
-            stable: Use stable mode
-            additional_args: Additional Arjun arguments
-
-        Returns:
-            HTTP parameter discovery results
-        """
-        data = {
-            "url": url,
-            "method": method,
-            "wordlist": wordlist,
-            "delay": delay,
-            "threads": threads,
-            "stable": stable,
-            "additional_args": additional_args
-        }
-        logger.info(f"🎯 Starting Arjun parameter discovery: {url}")
-        result = hexstrike_client.safe_post("api/tools/arjun", data)
-        if result.get("success"):
-            logger.info(f"✅ Arjun parameter discovery completed for {url}")
-        else:
-            logger.error(f"❌ Arjun parameter discovery failed for {url}")
-        return result
-
-    @mcp.tool()
-    def paramspider_mining(domain: str, level: int = 2,
-                          exclude: str = "png,jpg,gif,jpeg,swf,woff,svg,pdf,css,ico",
-                          output: str = "", additional_args: str = "") -> Dict[str, Any]:
-        """
-        Execute ParamSpider for parameter mining from web archives with enhanced logging.
-
-        Args:
-            domain: The target domain
-            level: Mining level depth
-            exclude: File extensions to exclude
-            output: Output file path
-            additional_args: Additional ParamSpider arguments
-
-        Returns:
-            Parameter mining results from web archives
-        """
-        data = {
-            "domain": domain,
-            "level": level,
-            "exclude": exclude,
-            "output": output,
-            "additional_args": additional_args
-        }
-        logger.info(f"🕷️  Starting ParamSpider mining: {domain}")
-        result = hexstrike_client.safe_post("api/tools/paramspider", data)
-        if result.get("success"):
-            logger.info(f"✅ ParamSpider mining completed for {domain}")
-        else:
-            logger.error(f"❌ ParamSpider mining failed for {domain}")
-        return result
-
     @mcp.tool()
     def x8_parameter_discovery(url: str, wordlist: str = "/usr/share/wordlists/x8/params.txt",
                               method: str = "GET", body: str = "", headers: str = "",
